VariationalInequalitySolver

Documentation for VariationalInequalitySolver.

VariationalInequalitySolver.hess_coord_residual

VariationalInequalitySolver.hess_coord_residual!

VariationalInequalitySolver.hess_op_residual

VariationalInequalitySolver.hess_op_residual!

VariationalInequalitySolver.hess_residual

VariationalInequalitySolver.hess_structure_residual

VariationalInequalitySolver.hess_structure_residual!

VariationalInequalitySolver.hprod_residual

VariationalInequalitySolver.hprod_residual!

VariationalInequalitySolver.jac_coord_residual

VariationalInequalitySolver.jac_coord_residual!

VariationalInequalitySolver.jac_op_residual

VariationalInequalitySolver.jac_op_residual!

VariationalInequalitySolver.jac_op_residual!

VariationalInequalitySolver.jac_op_residual!

VariationalInequalitySolver.jac_residual

VariationalInequalitySolver.jac_structure_residual

VariationalInequalitySolver.jac_structure_residual!

VariationalInequalitySolver.jprod_residual

VariationalInequalitySolver.jprod_residual!

VariationalInequalitySolver.jprod_residual!

VariationalInequalitySolver.jprod_residual!

VariationalInequalitySolver.jth_hess_residual

VariationalInequalitySolver.jtprod_residual

VariationalInequalitySolver.jtprod_residual!

VariationalInequalitySolver.jtprod_residual!

VariationalInequalitySolver.jtprod_residual!

VariationalInequalitySolver.project

VariationalInequalitySolver.project!

VariationalInequalitySolver.residual

VariationalInequalitySolver.residual!

VariationalInequalitySolver.hess_coord_residual! — Function

vals = hess_coord_residual!(model, x, v, vals)

Computes the linear combination of the Hessians of the residuals at x with coefficients v in sparse coordinate format, rewriting vals.

VariationalInequalitySolver.hess_coord_residual — Method

vals = hess_coord_residual(model, x, v)

Computes the linear combination of the Hessians of the residuals at x with coefficients v in sparse coordinate format.

VariationalInequalitySolver.hess_op_residual! — Method

Hop = hess_op_residual!(model, x, i, Hiv)

Computes the Hessian of the i-th residual at x, in linear operator form. The vector Hiv is used as preallocated storage for the operation.

VariationalInequalitySolver.hess_op_residual — Method

Hop = hess_op_residual(model, x, i)

Computes the Hessian of the i-th residual at x, in linear operator form.

VariationalInequalitySolver.hess_residual — Method

H = hess_residual(model, x, v)

Computes the linear combination of the Hessians of the residuals at x with coefficients v. A Symmetric object wrapping the lower triangle is returned.

VariationalInequalitySolver.hess_structure_residual! — Function

hess_structure_residual!(model, rows, cols)

Returns the structure of the residual Hessian in place.

VariationalInequalitySolver.hess_structure_residual — Method

(rows,cols) = hess_structure_residual(model)

Returns the structure of the residual Hessian.

VariationalInequalitySolver.hprod_residual! — Function

Hiv = hprod_residual!(model, x, i, v, Hiv)

Computes the product of the Hessian of the i-th residual at x, times the vector v, and stores it in vector Hiv.

VariationalInequalitySolver.hprod_residual — Method

Hiv = hprod_residual(model, x, i, v)

Computes the product of the Hessian of the i-th residual at x, times the vector v.

VariationalInequalitySolver.jac_coord_residual! — Function

vals = jac_coord_residual!(model, x, vals)

Computes the Jacobian of the residual at x in sparse coordinate format, rewriting vals. rows and cols are not rewritten.

VariationalInequalitySolver.jac_coord_residual — Method

(rows,cols,vals) = jac_coord_residual(model, x)

Computes the Jacobian of the residual at x in sparse coordinate format.

VariationalInequalitySolver.jac_op_residual! — Method

Jx = jac_op_residual!(model, x, Jv, Jtv)

Computes $J(x)$, the Jacobian of the residual at x, in linear operator form. The vectors Jv and Jtv are used as preallocated storage for the operations.

VariationalInequalitySolver.jac_op_residual! — Method

Jx = jac_op_residual!(model, x, rows, cols, Jv, Jtv)

Computes $J(x)$, the Jacobian of the residual at x, in linear operator form. The vectors Jv and Jtv are used as preallocated storage for the operations. The structure of the Jacobian should be given by (rows, cols).

VariationalInequalitySolver.jac_op_residual! — Method

Jx = jac_op_residual!(model, rows, cols, vals, Jv, Jtv)

Computes $J(x)$, the Jacobian of the residual given by (rows, cols, vals), in linear operator form. The vectors Jv and Jtv are used as preallocated storage for the operations.

VariationalInequalitySolver.jac_op_residual — Method

Jx = jac_op_residual(model, x)

Computes $J(x)$, the Jacobian of the residual at x, in linear operator form.

VariationalInequalitySolver.jac_residual — Method

Jx = jac_residual(model, x)

Computes $J(x)$, the Jacobian of the residual at x.

VariationalInequalitySolver.jac_structure_residual! — Function

(rows,cols) = jac_structure_residual!(model, rows, cols)

Returns the structure of the constraint's Jacobian in sparse coordinate format in place.

VariationalInequalitySolver.jac_structure_residual — Method

(rows,cols) = jac_structure_residual(model)

Returns the structure of the constraint's Jacobian in sparse coordinate format.

VariationalInequalitySolver.jprod_residual! — Function

Jv = jprod_residual!(model, x, v, Jv)

Computes the product of the Jacobian of the residual at x and a vector, i.e., $J(x)v$, storing it in Jv.

VariationalInequalitySolver.jprod_residual! — Method

Jv = jprod_residual!(model, x, rows, cols, v, Jv)

Computes the product of the Jacobian of the residual at x and a vector, i.e., $J(x)v$, storing it in Jv. The structure of the Jacobian is given by (rows, cols).

VariationalInequalitySolver.jprod_residual! — Method

Jv = jprod_residual!(model, rows, cols, vals, v, Jv)

Computes the product of the Jacobian of the residual given by (rows, cols, vals) and a vector, i.e., $J(x)v$, storing it in Jv.

VariationalInequalitySolver.jprod_residual — Method

Jv = jprod_residual(model, x, v)

Computes the product of the Jacobian of the residual at x and a vector, i.e., $J(x)v$.

VariationalInequalitySolver.jth_hess_residual — Method

Hj = jth_hess_residual(model, x, j)

Computes the Hessian of the j-th residual at x.

VariationalInequalitySolver.jtprod_residual! — Function

Jtv = jtprod_residual!(model, x, v, Jtv)

Computes the product of the transpose of the Jacobian of the residual at x and a vector, i.e., $J(x)^Tv$, storing it in Jtv.

VariationalInequalitySolver.jtprod_residual! — Method

Jtv = jtprod_residual!(model, x, rows, cols, v, Jtv)

Computes the product of the transpose Jacobian of the residual at x and a vector, i.e., $J(x)^Tv$, storing it in Jv. The structure of the Jacobian is given by (rows, cols).

VariationalInequalitySolver.jtprod_residual! — Method

Jtv = jtprod_residual!(model, rows, cols, vals, v, Jtv)

Computes the product of the transpose of the Jacobian of the residual given by (rows, cols, vals) and a vector, i.e., $J(x)^Tv$, storing it in Jv.

VariationalInequalitySolver.jtprod_residual — Method

Jtv = jtprod_residual(model, x, v)

Computes the product of the transpose of the Jacobian of the residual at x and a vector, i.e., $J(x)^Tv$.

VariationalInequalitySolver.project! — Function

project!(Px::AbstractVector{T}, model::VIModel, x::AbstractVector{T}) where {T}

Compute the projection of d over X in-place, i.e.,

\[ min_x 0.5 | d - x |²₂ s.t. x ∈ X\]

VariationalInequalitySolver.project — Method

project(model::VIModel, x::AbstractVector{T}) where {T}

Compute the projection of d over X, i.e.,

\[ min_x 0.5 | d - x |²₂ s.t. x ∈ X\]

```

VariationalInequalitySolver.residual! — Function

Fx = residual!(model, x, Fx)

Computes $F(x)$, the residual at x.

VariationalInequalitySolver.residual — Method

Fx = residual(model, x)

Computes $F(x)$, the residual at x.