VariationalInequalitySolver

Documentation for VariationalInequalitySolver.

NLPModels.hess_coord_residual

NLPModels.hess_coord_residual!

NLPModels.hess_op_residual

NLPModels.hess_residual

NLPModels.hess_structure_residual

NLPModels.hess_structure_residual!

NLPModels.hprod_residual

NLPModels.hprod_residual!

NLPModels.jac_coord_residual

NLPModels.jac_coord_residual!

NLPModels.jac_op_residual

NLPModels.jac_op_residual!

NLPModels.jac_residual

NLPModels.jac_structure_residual

NLPModels.jac_structure_residual!

NLPModels.jprod_residual

NLPModels.jprod_residual!

NLPModels.jprod_residual!

NLPModels.jth_hess_residual

NLPModels.jtprod_residual

NLPModels.jtprod_residual!

NLPModels.jtprod_residual!

NLPModels.residual

NLPModels.residual!

VariationalInequalitySolver.ProjectedReflectedGradientVI

VariationalInequalitySolver.project

VariationalInequalitySolver.project!

NLPModels.hess_coord_residual! — Function

vals = hess_coord_residual!(model, x, v, vals)

Computes the linear combination of the Hessians of the residuals at x with coefficients v in sparse coordinate format, rewriting vals.

NLPModels.hess_coord_residual — Method

vals = hess_coord_residual(model, x, v)

Computes the linear combination of the Hessians of the residuals at x with coefficients v in sparse coordinate format.

NLPModels.hess_op_residual — Method

Hop = hess_op_residual(model, x, i)

Computes the Hessian of the i-th residual at x, in linear operator form.

NLPModels.hess_residual — Method

H = hess_residual(model, x, v)

Computes the linear combination of the Hessians of the residuals at x with coefficients v. A Symmetric object wrapping the lower triangle is returned.

NLPModels.hess_structure_residual! — Function

hess_structure_residual!(model, rows, cols)

Returns the structure of the residual Hessian in place.

NLPModels.hess_structure_residual — Method

(rows,cols) = hess_structure_residual(model)

Returns the structure of the residual Hessian.

NLPModels.hprod_residual! — Function

Hiv = hprod_residual!(model, x, i, v, Hiv)

Computes the product of the Hessian of the i-th residual at x, times the vector v, and stores it in vector Hiv.

NLPModels.hprod_residual — Method

Hiv = hprod_residual(model, x, i, v)

Computes the product of the Hessian of the i-th residual at x, times the vector v.

NLPModels.jac_coord_residual! — Function

vals = jac_coord_residual!(model, x, vals)

Computes the Jacobian of the residual at x in sparse coordinate format, rewriting vals. rows and cols are not rewritten.

NLPModels.jac_coord_residual — Method

(rows,cols,vals) = jac_coord_residual(model, x)

Computes the Jacobian of the residual at x in sparse coordinate format.

NLPModels.jac_op_residual! — Method

Jx = jac_op_residual!(model, rows, cols, vals, Jv, Jtv)

Computes $J(x)$, the Jacobian of the residual given by (rows, cols, vals), in linear operator form. The vectors Jv and Jtv are used as preallocated storage for the operations.

NLPModels.jac_op_residual — Method

Jx = jac_op_residual(model, x)

Computes $J(x)$, the Jacobian of the residual at x, in linear operator form.

NLPModels.jac_residual — Method

Jx = jac_residual(model, x)

Computes $J(x)$, the Jacobian of the residual at x.

NLPModels.jac_structure_residual! — Function

(rows,cols) = jac_structure_residual!(model, rows, cols)

Returns the structure of the constraint's Jacobian in sparse coordinate format in place.

NLPModels.jac_structure_residual — Method

(rows,cols) = jac_structure_residual(model)

Returns the structure of the constraint's Jacobian in sparse coordinate format.

NLPModels.jprod_residual! — Function

Jv = jprod_residual!(model, x, v, Jv)

Computes the product of the Jacobian of the residual at x and a vector, i.e., $J(x)v$, storing it in Jv.

NLPModels.jprod_residual! — Method

Jv = jprod_residual!(model, rows, cols, vals, v, Jv)

Computes the product of the Jacobian of the residual given by (rows, cols, vals) and a vector, i.e., $J(x)v$, storing it in Jv.

NLPModels.jprod_residual — Method

Jv = jprod_residual(model, x, v)

Computes the product of the Jacobian of the residual at x and a vector, i.e., $J(x)v$.

NLPModels.jth_hess_residual — Method

Hj = jth_hess_residual(model, x, j)

Computes the Hessian of the j-th residual at x.

NLPModels.jtprod_residual! — Function

Jtv = jtprod_residual!(model, x, v, Jtv)

Computes the product of the transpose of the Jacobian of the residual at x and a vector, i.e., $J(x)^Tv$, storing it in Jtv.

NLPModels.jtprod_residual! — Method

Jtv = jtprod_residual!(model, rows, cols, vals, v, Jtv)

Computes the product of the transpose of the Jacobian of the residual given by (rows, cols, vals) and a vector, i.e., $J(x)^Tv$, storing it in Jv.

NLPModels.jtprod_residual — Method

Jtv = jtprod_residual(model, x, v)

Computes the product of the transpose of the Jacobian of the residual at x and a vector, i.e., $J(x)^Tv$.

NLPModels.residual! — Function

Fx = residual!(model, x, Fx)

Computes $F(x)$, the residual at x.

NLPModels.residual — Method

Fx = residual(model, x)

Computes $F(x)$, the residual at x.

VariationalInequalitySolver.ProjectedReflectedGradientVI — Method

Yura Malitsky. Projected reflected gradient methods for monotone variational inequalities. SIAM Journal on Optimization, 25(1):502–520, 2015.

VariationalInequalitySolver.project! — Function

project!(Px::AbstractVector{T}, model::VIModel, x::AbstractVector{T}) where {T}

Compute the projection of d over X in-place, i.e.,

\[ min_x 0.5 | d - x |²₂ s.t. x ∈ X\]

VariationalInequalitySolver.project — Method

project(model::VIModel, x::AbstractVector{T}) where {T}

Compute the projection of d over X, i.e.,

\[ min_x 0.5 | d - x |²₂ s.t. x ∈ X\]

```